math cleanup
Former-commit-id: 606539b05aff0d93e63ba7602171a66e55dc53b7

+++ ork.core/src/math/PIDController.cpp
@@ -96,7 +96,7 @@ template<typename T>
 T PIDController2<T>::Update(T MeasuredError, float dt )
 {
     mIntegral += MeasuredError*dt;
-	mIntegral *= std::pow(mIntegralDecay, dt);
+	mIntegral *= powf(mIntegralDecay, dt);
 
 	//mIntegral = maximum(mIntegral, mMaxDelta.GetX());
 	//mIntegral = minimum(mIntegral, mMaxDelta.GetY());
