+++ ork.lev2/src/aud/singularity/fft.cpp
@@ -101,7 +101,7 @@ namespace audiofft
       {
         if (_size != size)
         {
-          _ip.resize(2 + static_cast<int>(std::sqrt(static_cast<double>(size))));
+          _ip.resize(2 + static_cast<int>(sqrtf(static_cast<double>(size))));
           _w.resize(size / 2);
           _buffer.resize(size);
           _size = size;
